implement basic support for CSS background-image
Elements with [data-slimmage-bg] are handled similarly to
img[data-slimmage], but the resulting image url is set as their CSS
background-image property.

This is meant to be used for background images that are 100% width, i.e.
`background-size: 100% auto` is recommended. If you don't use that,
you get slightly cropped images on normal displays (because resolution
is rounded up in 160px increments by default) and blocky half images on
high-dpi displays.

The implementation requires querySelectorAll (IE8+). Note that
background-size doesn't work on IE8 so check the usual workarounds for
that.